Hotfix gitbox syncing process
Seems like GitHub breaks on occasion which breaks the repos. This should hotfix the issue by retrying up to five times until we either get a success or we give up. There's an issue with PR merges only transferring up to 20 at a time, which may be what's happening here...